fixed memory

+++ kernel/src/memory/malloc.rs
@@ -8,7 +8,7 @@ pub fn malloc(size: usize) -> *mut u8 {
         let result = CURR_PTR;
         CURR_PTR = CURR_PTR.add(size);
         while (CURR_PAGE as u64) < (CURR_PTR as u64) {
-            map_page(CURR_PAGE as u64, find_free_page() as u64, true, false);
+            map_page(CURR_PAGE, find_free_page(), true, false);
             CURR_PAGE = CURR_PAGE.add(PAGE_SIZE as usize);
         }
         result
